Change sql schema format to :sql
This will generate the file db/structure.sql. The file db/schema.rb won't be used anymore. This change was applied because when adding a enum type to postgress the schema was not generated propertly and the step `db:test:load` didn't succeed while running tests.

In the schema this error was shown as a comment "Unknown type xxxx for column xxx"

 support enum types in the test environment. In development was fine but the schema wasn't generated so the comand `db:test:load` never succeeded.  The shown error was "rails schema  This error was shown as a comment in the schema.rb